+++ cli-tool/test/commands/export.rb
@@ -31,7 +31,7 @@ class TestCommandsExport < MTest::Unit::TestCase
   end
 
   def test_create_build_cache_folder
-    Taylor::Commands::Export.new(['--dry-run'], { build_cache: '/tmp/build/' })
+    Taylor::Commands::Export.new(['--dry-run'], { 'build_cache' => '/tmp/build/' })
     assert_false Dir.exists?('./exports')
   end
 
@@ -46,7 +46,7 @@ class TestCommandsExport < MTest::Unit::TestCase
   end
 
   def test_check_docker_command_override_export_directory
-    export_command = Taylor::Commands::Export.new(['--dry-run'], { export_directory: '/tmp/exports' })
+    export_command = Taylor::Commands::Export.new(['--dry-run'], { 'export_directory' => '/tmp/exports' })
     assert_include export_command.puts_data, 'docker run'
     assert_include export_command.puts_data, '/tmp/exports'
     assert_include export_command.puts_data, "hellrok/taylor:linux-v#{TAYLOR_VERSION}"
@@ -56,7 +56,7 @@ class TestCommandsExport < MTest::Unit::TestCase
   end
 
   def test_check_docker_command_override_targets
-    export_command = Taylor::Commands::Export.new(['--dry-run'], { export_targets: 'linux,web' })
+    export_command = Taylor::Commands::Export.new(['--dry-run'], { 'export_targets' => 'linux,web' })
     assert_include export_command.puts_data, 'docker run'
     assert_include export_command.puts_data, File.join(Dir.pwd, 'exports')
     assert_include export_command.puts_data, "hellrok/taylor:linux-v#{TAYLOR_VERSION}"