+++ test/fixtures/models.rb
@@ -1,2 +1,2 @@
-require 'test/fixtures/models/camera2d'
-require 'test/fixtures/models/image'
+require 'fixtures/models/camera2d'
+require 'fixtures/models/image'