+++ test/images.rb
@@ -1,6 +1,6 @@
 class TestImage < MTest::Unit::TestCase
   def test_image_load
-    image = load_image('test/assets/test.png')
+    image = load_image('assets/test.png')
     assert_equal fixture_image_load, image.data
     unload_image(image)
   end
@@ -20,7 +20,7 @@ class TestImage < MTest::Unit::TestCase
   end
 
   def test_image_from_image
-    image = load_image('test/assets/test.png')
+    image = load_image('assets/test.png')
     new_image = image_from_image(image, Rectangle.new(1, 1, 2, 2))
     assert_equal fixture_image_from_image, new_image.data
     unload_image(image)
@@ -31,7 +31,7 @@ class TestImage < MTest::Unit::TestCase
     skip_unless_display_present
 
     set_window_title(__method__.to_s)
-    font = Font.load('./test/assets/tiny.ttf', size: 16)
+    font = Font.load('./assets/tiny.ttf', size: 16)
 
     image = image_text_ex(font, 'S', 16, 0, BLACK)
     assert_equal fixture_image_text_ex, image.data
@@ -41,29 +41,29 @@ class TestImage < MTest::Unit::TestCase
   end
 
   def test_image_resize!
-    image = load_image('test/assets/test.png')
+    image = load_image('assets/test.png')
     image_resize!(image, 6, 6)
     assert_equal fixture_image_resize!, image.data
     unload_image(image)
   end
 
   def test_image_resize_nearest_neighbour!
-    image = load_image('test/assets/test.png')
+    image = load_image('assets/test.png')
     image_resize_nearest_neighbour!(image, 6, 6)
     assert_equal fixture_image_resize_nearest_neighbour!, image.data
     unload_image(image)
   end
 
   def test_image_crop!
-    image = load_image('test/assets/test.png')
+    image = load_image('assets/test.png')
     image_crop!(image, Rectangle.new(0, 0, 3, 2))
     assert_equal fixture_image_crop!, image.data
     unload_image(image)
   end
 
   def test_image_alpha_mask!
-    image = load_image('test/assets/test.png')
-    mask = load_image('test/assets/alpha.png')
+    image = load_image('assets/test.png')
+    mask = load_image('assets/alpha.png')
 
     image_alpha_mask!(image, mask)
     assert_equal fixture_image_alpha_mask!, image.data
@@ -73,7 +73,7 @@ class TestImage < MTest::Unit::TestCase
   end
 
   def test_image_mipmaps!
-    image = load_image('test/assets/test.png')
+    image = load_image('assets/test.png')
     assert_equal 1, image.mipmaps
 
     image_mipmaps!(image)
@@ -83,7 +83,7 @@ class TestImage < MTest::Unit::TestCase
   end
 
   def test_image_flip_vertical!
-    image = load_image('test/assets/asymettrical.png')
+    image = load_image('assets/asymettrical.png')
 
     image_flip_vertical!(image)
     assert_equal fixture_image_flip_vertical!, image.data
@@ -92,7 +92,7 @@ class TestImage < MTest::Unit::TestCase
   end
 
   def test_image_flip_horizontal!
-    image = load_image('test/assets/asymettrical.png')
+    image = load_image('assets/asymettrical.png')
 
     image_flip_horizontal!(image)
     assert_equal fixture_image_flip_horizontal!, image.data
@@ -101,7 +101,7 @@ class TestImage < MTest::Unit::TestCase
   end
 
   def test_image_rotate_cw!
-    image = load_image('test/assets/asymettrical.png')
+    image = load_image('assets/asymettrical.png')
 
     image_rotate_cw!(image)
     assert_equal fixture_image_rotate_cw!, image.data
@@ -110,7 +110,7 @@ class TestImage < MTest::Unit::TestCase
   end
 
   def test_image_rotate_ccw!
-    image = load_image('test/assets/asymettrical.png')
+    image = load_image('assets/asymettrical.png')
 
     image_rotate_ccw!(image)
     assert_equal fixture_image_rotate_ccw!, image.data
@@ -183,7 +183,7 @@ class TestImage < MTest::Unit::TestCase
   end
 
   def test_image_colour_replace!
-    image = load_image('test/assets/test.png')
+    image = load_image('assets/test.png')
 
     image_colour_replace!(image, WHITE, BLUE)
     assert_equal fixture_image_colour_replace!, image.data