+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Taylor
 
+## Unreleased
+
 ## v0.3.3
 
 - Add regex support