+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Add Gitpod support
 - Add more tests for cli-export
+- Use slim docker images to save a few megabytes
+- Building the export docker images now includes the version for easy use
 
 ## v0.3.3