+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Building the export docker images now includes the version for easy use
 - cli-export run now properly handles being passed a path
 - Test suite is now a proper Taylor application
+- Test suite now mostly runs on Windows
 
 ## v0.3.3