+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Taylor
 
-## Unreleased
+## v0.3.4
 
 - Add Gitpod support
 - Add more tests for cli-export